refcatoring to fit standards

+++ darwin/options.py
@@ -217,10 +217,10 @@ class Options(object):
         )
         parser_comment.add_argument("file", type=str, help="File to comment")
         parser_comment.add_argument("--comment-text", type=str, help="Comment: list of words")
-        parser_comment.add_argument("--comment-x-position", type=int, default=1, help="X coordinate for comment")
-        parser_comment.add_argument("--comment-y-position", type=int, default=1, help="Y coordinate for comment")
-        parser_comment.add_argument("--comment-width-px", type=int, default=1, help="Comment width in pixels")
-        parser_comment.add_argument("--comment-height-px", type=int, default=1, help="Comment height in pixels")
+        parser_comment.add_argument("--comment-x-position", type=float, default=1, help="X coordinate for comment")
+        parser_comment.add_argument("--comment-y-position", type=float, default=1, help="Y coordinate for comment")
+        parser_comment.add_argument("--comment-width-px", type=float, default=1, help="Comment width in pixels")
+        parser_comment.add_argument("--comment-height-px", type=float, default=1, help="Comment height in pixels")
 
         # Help
         dataset_action.add_parser("help", help="Show this help message and exit.")
